+++ src/test/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/validators/OcrDataValidatorTest.java
@@ -12,22 +12,22 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.ADDRESS_LINE_1;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.ADDRESS_LINE_2;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.ADDRESS_LINE_3;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.CONTACT_NUMBER;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.COUNTRY;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.COUNTY;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.DATE_OF_BIRTH;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.EMAIL;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.FIRST_NAME;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.LAST_NAME;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.POST_CODE;
+import static uk.gov.hmcts.reform.bulkscanccdeventhandler.common.OcrFieldNames.POST_TOWN;
 import static uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.FormType.CONTACT;
 import static uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.FormType.PERSONAL;
 import static uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.out.ValidationStatus.ERRORS;
 import static uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.out.ValidationStatus.SUCCESS;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.ADDRESS_LINE_1;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.ADDRESS_LINE_2;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.ADDRESS_LINE_3;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.CONTACT_NUMBER;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.COUNTRY;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.COUNTY;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.DATE_OF_BIRTH;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.EMAIL;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.FIRST_NAME;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.LAST_NAME;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.POST_CODE;
-import static uk.gov.hmcts.reform.bulkscanccdeventhandler.shared.OcrFieldNames.POST_TOWN;
 
 class OcrDataValidatorTest {
